Change inheritance structure of animation
To allow for appending a Sequence to a Sequence, a new class
called AnimationBase was added to allow for all three types of
Animations (Interval, Animation and Sequence) to be passed to
the `append` method.

Changed status when a conflict occurs to 4 to be able to know
when a conflict actually happened.